requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
beautifulsoup4==4.11.1
bleach==5.0.0
certifi==2022.6.15
chardet==4.0.0
charset-normalizer==2.0.12
cycler==0.11.0
dataframe-image==0.1.1
defusedxml==0.7.1
discord.py==1.7.3
entrypoints==0.4
fastjsonschema==2.15.3
fonttools==4.33.3
idna==3.3
Jinja2==3.1.2
jsonschema==4.6.0
jupyter-client==7.3.4
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
kiwisolver==1.4.3
MarkupSafe==2.1.1
matplotlib==3.5.2
mistune==0.8.4
multidict==6.0.2
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
numpy==1.22.4
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
Pillow==9.1.1
Pygments==2.12.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pyzmq==23.2.0
requests==2.28.0
six==1.16.0
soupsieve==2.3.2.post1
tinycss2==1.1.1
tornado==6.1
traitlets==5.3.0
typing_extensions==4.2.0
urllib3==1.26.9
webencodings==0.5.1
yarl==1.7.2
